+++ packages/create-monokle-plugin/template-validation-ts/src/rules/1-example.ts
@@ -1,4 +1,4 @@
-import { defineRule } from "@monokle/validation/custom";
+import { defineRule } from "@monokle/plugin-toolkit";
 
 /**
  * A basic example.
@@ -9,10 +9,8 @@ export const noEmptyAnnotations = defineRule({
   help: "Add any annotation to the Kubernetes resource.",
   validate({ resources }, { report }) {
     resources.forEach((resource) => {
-      const annotations = Object.entries(
-        resource.content.metadata?.annotations ?? {}
-      );
-      const hasAnnotations = annotations.length > 1;
+      const annotations = Object.entries(resource.metadata?.annotations ?? {});
+      const hasAnnotations = annotations.length > 0;
 
       if (!hasAnnotations) {
         report(resource, { path: "metadata.annotations" });