+++ packages/validation/src/validators/resource-links/validator.ts
@@ -1,11 +1,6 @@
 import { AbstractPlugin } from "../../common/AbstractPlugin.js";
-import { ValidationResult, Region } from "../../common/sarif.js";
-import {
-  Resource,
-  ResourceRef,
-  ResourceRefType,
-  ToolConfig,
-} from "../../common/types.js";
+import { Region, ValidationResult } from "../../common/sarif.js";
+import { Resource, ResourceRef, ResourceRefType, ToolConfig } from "../../common/types.js";
 import { createLocations } from "../../utils/createLocations.js";
 import { isDefined } from "../../utils/isDefined.js";
 import { RESOURCE_LINK_RULES } from "./rules.js";
@@ -24,9 +19,9 @@ export class ResourceLinksValidator extends AbstractPlugin {
         id: "LNK",
         name: "resource-links",
         displayName: "Resource Links",
-        description: "Validates that references to other resources are valid.",
+        description: "Validates that links to other resources are valid.",
         icon: "resource-links",
-        learnMoreUrl: "https://kubeshop.github.io/monokle/resource-validation/",
+        learnMoreUrl: "https://kubeshop.github.io/monokle/resource-validation/"
       },
       RESOURCE_LINK_RULES
     );
@@ -56,7 +51,7 @@ export class ResourceLinksValidator extends AbstractPlugin {
     resource: Resource
   ): Promise<ValidationResult[]> {
     const refs = resource.refs ?? [];
-    const unsatisfiedRefs = refs.filter(isUnsatisfied);
+    const unsatisfiedRefs = refs.filter(ref => ref.type === ResourceRefType.Unsatisfied );
 
     const results = unsatisfiedRefs
       .map((ref) => this.adaptToValidationResult(resource, ref))
@@ -70,28 +65,37 @@ export class ResourceLinksValidator extends AbstractPlugin {
     const region: Region | undefined =
       pos?.endColumn && pos.endLine
         ? {
-            startLine: pos.line,
-            startColumn: pos.column,
-            endLine: pos.endLine,
-            endColumn: pos.endColumn,
-          }
+          startLine: pos.line,
+          startColumn: pos.column,
+          endLine: pos.endLine,
+          endColumn: pos.endColumn
+        }
         : pos
-        ? {
+          ? {
             startLine: pos.line,
             startColumn: pos.column,
             endLine: pos.line,
-            endColumn: pos.column + pos.length,
+            endColumn: pos.column + pos.length
           }
-        : undefined;
+          : undefined;
 
     const locations = createLocations(resource, region);
 
-    return this.createValidationResult("LNK001", {
-      message: {
-        text: "Unsatisfied resource link.",
-      },
-      locations,
-    });
+    if (ref.target?.type === "resource" && ref.target?.isOptional) {
+      return this.isRuleEnabled("LNK002") ? this.createValidationResult("LNK002", {
+        message: {
+          text: "Unsatisfied optional resource link."
+        },
+        locations
+      }) : undefined;
+    } else {
+      return this.isRuleEnabled("LNK001") ? this.createValidationResult("LNK001", {
+        message: {
+          text: "Unsatisfied resource link."
+        },
+        locations
+      }) : undefined;
+    }
   }
 }