+++ packages/create-monokle-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # create-monokle-plugin
 
+## 0.3.1
+
+### Patch Changes
+
+- 4a7a899: Add readme extension
+- d1e0ea0: Delete internals
+- 5f27e51: Fix gitignore
+
+## 0.3.0
+
+### Minor Changes
+
+- ab4d567: Add template with toolkit and fix minor bug
+
 ## 0.2.0
 
 ### Minor Changes